# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/activate/device/activation-file/{sn}": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "device"
                ],
                "summary": "获取设备激活文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "产品ID",
                        "name": "productID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "设备序列号",
                        "name": "sn",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "激活文件",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/activate/device/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "添加单个设备",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "参数：添加设备",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeviceAdd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加设备",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/activate/device/batch-add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "批量添加设备",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "参数：批量添加设备",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeviceBatchAdd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "批量添加设备",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/activate/device/batch-update-license": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "批量更新设备许可证类型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "参数：批量更新许可证类型",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeviceBatchUpdateLicense"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "批量更新许可证类型",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/activate/device/license-types": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "获取产品下的许可证类型列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "产品ID",
                        "name": "product_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取许可证类型列表",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/activate/device/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "获取产品下的设备列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "产品ID",
                        "name": "product_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "许可证类型ID",
                        "name": "license_type_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "设备序列号",
                        "name": "sn",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，从1开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取设备列表",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/activate/device/products": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "获取产品列表（带设备数量）",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，从1开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取产品列表",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/activate/device/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "通过SN获取设备信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "设备序列号",
                        "name": "sn",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取设备信息",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/activate/device/update": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "更新设备",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "参数：更新设备",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeviceUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新设备",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/activate/device/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "删除设备",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "设备ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除设备",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.DeviceAdd": {
            "type": "object",
            "required": [
                "license_type_id",
                "product_id",
                "sn"
            ],
            "properties": {
                "license_type_id": {
                    "type": "integer"
                },
                "oem_tag": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "sn": {
                    "type": "string"
                }
            }
        },
        "dto.DeviceBatchAdd": {
            "type": "object",
            "required": [
                "license_type_id",
                "product_id",
                "sns"
            ],
            "properties": {
                "license_type_id": {
                    "type": "integer"
                },
                "oem_tag": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "sns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.DeviceBatchUpdateLicense": {
            "type": "object",
            "required": [
                "device_ids",
                "license_type_id"
            ],
            "properties": {
                "device_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "license_type_id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
        "dto.DeviceUpdate": {
            "type": "object",
            "required": [
                "id",
                "license_type_id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "license_type_id": {
                    "type": "integer"
                },
                "oem_tag": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
...
        "resp.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}